requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
defusedxml==0.6.0
Django==3.1.9
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-storages==1.10.1
idna==2.10
importlib-metadata==1.7.0
jeepney==0.4.3
jmespath==0.10.0
jsonconfig-tool==1.5
keyring==21.3.0
messages==0.5.0
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
python-box==5.1.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
validus==0.3.0
zipp==3.1.0
